requirements.txt
flask==2.0.1
requests==2.28.1
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
python-dotenv==1.0.0 